Use FAB vom design lib instead of own creation
