+++ ShareViaHttp/app/src/main/res/layout/fab_with_text.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+       xmlns:app="http://schemas.android.com/apk/res-auto"
+       xmlns:tools="http://schemas.android.com/tools"
+       android:layout_width="match_parent"
+       android:layout_height="match_parent">
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fab_with_text_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:clickable="false"
+        app:elevation="2dp"
+        app:fabSize="mini"
+        tools:src="@drawable/ic_qrcode"/>
+
+    <TextView
+        android:id="@+id/fab_with_text_text"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/fab_with_text_fab"
+        android:layout_centerInParent="true"
+        tools:text="ExampleText"/>
+
+</merge>
