+++ ShareViaHttp/app/src/main/res/layout/navigation_layout.xml
@@ -9,28 +9,28 @@
     android:gravity="center_horizontal"
     android:orientation="horizontal">
 
-    <com.MarcosDiez.shareviahttp.views.RoundedImageButtonWithTextView
+    <com.MarcosDiez.shareviahttp.views.FloatingActionButtonWithText
         android:id="@+id/button_qr_code"
         android:layout_width="90dp"
         android:layout_height="95dp"
         app:imageSrc="@drawable/ic_qrcode"
         app:text="@string/button_qrcode" />
 
-    <com.MarcosDiez.shareviahttp.views.RoundedImageButtonWithTextView
+    <com.MarcosDiez.shareviahttp.views.FloatingActionButtonWithText
         android:id="@+id/change_ip"
         android:layout_width="90dp"
         android:layout_height="95dp"
         app:imageSrc="@drawable/ic_ip"
         app:text="@string/change_ip" />
 
-    <com.MarcosDiez.shareviahttp.views.RoundedImageButtonWithTextView
+    <com.MarcosDiez.shareviahttp.views.FloatingActionButtonWithText
         android:id="@+id/stop_server"
         android:layout_width="90dp"
         android:layout_height="95dp"
         app:imageSrc="@drawable/ic_stop"
         app:text="@string/stop_server" />
 
-    <com.MarcosDiez.shareviahttp.views.RoundedImageButtonWithTextView
+    <com.MarcosDiez.shareviahttp.views.FloatingActionButtonWithText
         android:id="@+id/button_share_url"
         android:layout_width="90dp"
         android:layout_height="95dp"
